Refactor unpackThreeModel
