+++ js/scripts/templates/js_wrapper.mustache
@@ -6,8 +6,7 @@ var _ = require('underscore');
 var THREE = require('three');
 var widgets = require('@jupyter-widgets/base');
 var dataserializers = require('jupyter-dataserializers');
-
-var unpackThreeModel = require('{{ relativePathToBase }}/_base/Three').unpackThreeModel;
+var serializers = require('{{ relativePathToBase }}/_base/serializers');
 
 var {{ superClass.modelName }} = require('{{ superClass.requirePath }}').{{ superClass.modelName }};
