+++ js/src/renderers/WebGLRenderer.js
@@ -8,7 +8,7 @@ var Promise = require('bluebird');
 
 var RenderableModel = require('../_base/Renderable').RenderableModel;
 var RenderableView = require('../_base/Renderable').RenderableView;
-var unpackThreeModel = require('../_base/Three').unpackThreeModel;
+var unpackThreeModel = require('../_base/serializers').unpackThreeModel;
 
 
 var WebGLRendererModel = RenderableModel.extend({
